streamlit_app/requirements.txt
streamlit>=1.28.0
tensorflow>=2.15.0
tf-keras>=2.15.0
pillow>=9.5.0
numpy>=1.23.0
